+++ src/index.ts
@@ -1,16 +1,18 @@
-import constant from './constant'
-import freeze from './freeze'
-import is from './is'
-import _if from './if'
-import ifElse from './ifElse'
-import map from './map'
-import omit from './omit'
-import omitBy from './omitBy'
-import reject from './reject'
-import update, {omitted} from './update'
-import updateIn from './updateIn'
-import withDefault from './withDefault'
-import {_} from './util/curry'
+import constant from "./constant";
+import freeze from "./freeze";
+import is from "./is";
+import _if from "./if";
+import ifElse from "./ifElse";
+import map from "./map";
+import omit from "./omit";
+import omitBy from "./omitBy";
+import reject from "./reject";
+import update, { omitted } from "./update";
+import updateIn from "./updateIn";
+import withDefault from "./withDefault";
+import { _ } from "./util/curry";
+import mapWhen from "./mapWhen";
+import mapWhenElse from "./mapWhenElse";
 
 const methods = {
   constant,
@@ -26,12 +28,14 @@ const methods = {
   reject,
   omitted,
   _,
-}
+  mapWhen,
+  mapWhenElse,
+};
 
-export type Updeep = typeof update & typeof methods
+export type Updeep = typeof update & typeof methods;
 
-const u: Updeep = update as Updeep
+const u: Updeep = update as Updeep;
 
-Object.assign(u, methods)
+Object.assign(u, methods);
 
-export default u
+export default u;